refactor(smb): split the handler god file by state concern
